requirements.txt
# requirements.txt
ollama
pandas
networkx
pyvis
python-dotenv
tiktoken
numpy
tqdm